redesign template engine
